+++ app/login/page.tsx
@@ -8,6 +8,7 @@ import { ArrowRight, Eye, EyeOff } from 'lucide-react'
 
 export default function LoginInterface() {
   const [username, setUsername] = useState('')
+  const [error, setError] = useState("")
   const [password, setPassword] = useState('')
   const [showPassword, setShowPassword] = useState(false)
   const [isLoading, setIsLoading] = useState(false)
@@ -15,14 +16,36 @@ export default function LoginInterface() {
 
   const handleContinue = async () => {
     if (step === 'username' && username) {
-      setStep('password')
-    } else if (step === 'password' && password) {
-      setIsLoading(true)
-      // Simulate login
-      setTimeout(() => {
-        setIsLoading(false)
-        // Handle login success
-      }, 1500)
+      setStep("password")
+    } else if (step === "password" && password) {
+      try {
+        setIsLoading(true);
+        const res = await fetch("/api/login", {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({ username, password })
+        });
+        if (res.ok) {
+          location.href = "/";
+        } else if (res.status === 404) {
+          const reg = await fetch("/api/register", {
+            method: "POST",
+            headers: { "Content-Type": "application/json" },
+            body: JSON.stringify({ username, password })
+          });
+          if (reg.ok) {
+            location.href = "/";
+          } else {
+            const data = await reg.json();
+            setError(data.error || "Unable to register");
+          }
+        } else {
+          const data = await res.json();
+          setError(data.error || "Invalid credentials");
+        }
+      } finally {
+        setIsLoading(false);
+      }
     }
   }
 
@@ -163,6 +186,7 @@ export default function LoginInterface() {
                 </span>
               )}
             </Button>
+            {error && <p className="text-sm text-red-500 mt-2">{error}</p>}
           </motion.div>
         </div>
